ci: make workflows better work for PR created from forked repo (#225)
"contribution checks"
The workflow now runs on `pull_request_target` events.
There are no security issues here. Checks are made only on the updated
PR file without doing any tool installation, cache update or branch
check. Only the GitHub API is used.
Using this event allows you to create a PR comment when the PR is
created from a forked repository.

"build preview" and "references validation" workflows.
The content of the branch of the fork is now correctly used. Previously,
the branch of the fork wasn't found by Antora, so the content of the
generated site was empty.

### Notes

Covers
https://github.com/bonitasoft/bonita-documentation-site/issues/402
Covers
https://github.com/bonitasoft/bonita-documentation-site/issues/685